Update paths to ignore in CI for Grafana Container
This commit updates the paths to ignore in the CI for the Grafana Container. The paths for '.gitignore', '*.md', 'LICENSE', 'install_docker_debian.sh', 'install_docker_ubuntu.sh', and 'certs/*' are now ignored during push and pull requests. This prevents unnecessary CI runs when changes are made to these files.